x32 port (diff against vanilla x86_64)

+++ arch/x32/bits/ipc.h
@@ -7,8 +7,8 @@ struct ipc_perm
 	gid_t cgid;
 	mode_t mode;
 	int __ipc_perm_seq;
-	long __pad1;
-	long __pad2;
+	long long __pad1;
+	long long __pad2;
 };
 
 #define IPC_64 0
