+++ arch/x32/bits/setjmp.h
@@ -1,1 +1,1 @@
-typedef unsigned long __jmp_buf[8];
+typedef unsigned long long __jmp_buf[8];
